scripts/us_bea/states_gdp/requirements.txt
absl-py>=0.9.0
numpy>=1.18.5
pandas>=1.0.4
requests>=2.23.0
